Refactor to track path instead of visited nodes
Removed visited nodes and instead changed the queue objects to include
the node and path as the object, and then appended to the path
when queueing new things.

When a match is found, return the path string. Might return array of
arrays instead at some point depending on what is needed, but a string
works for now.